requirements.txt
Django>=4.2
djangorestframework>=3.16
django-cors-headers>=4.0
django_cleanup>=9.0.0
